Better fix for processIntent, moving method into SwrveAdmPushSupport
