+++ native/android/SwrveSDKPushSupport/src/amazon/java/com/swrve/unity/adm/SwrveAdmPushSupport.java
@@ -6,7 +6,9 @@ import java.util.List;
 
 import android.app.Activity;
 import android.content.Context;
+import android.content.Intent;
 import android.content.SharedPreferences;
+import android.os.Bundle;
 import android.util.Log;
 
 import com.amazon.device.messaging.ADM;
@@ -196,6 +198,24 @@ public class SwrveAdmPushSupport {
                     it.remove();
                 }
             }
+        }
+    }
+
+    protected static void processIntent(Context context, Intent intent) {
+        if (intent == null) {
+            return;
+        }
+        try {
+            Bundle extras = intent.getExtras();
+            if (extras != null && !extras.isEmpty()) {
+                Bundle msg = extras.getBundle("notification");
+                if (msg != null) {
+                    SwrveNotification notification = SwrveNotification.Builder.build(msg);
+                    SwrveAdmPushSupport.newOpenedNotification(context, notification);
+                }
+            }
+        } catch(Exception ex) {
+            Log.e(TAG, "Could not process push notification intent", ex);
         }
     }
 }
